FIX: move tuple_handler function to tools.py
